+++ SpruceCube/app/src/main/java/com/sprucecube/homeautomation/DevicesActivity.java
@@ -7,6 +7,7 @@ import android.support.annotation.NonNull;
 import android.support.design.widget.BottomNavigationView;
 import android.support.design.widget.NavigationView;
 import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
 import android.support.v4.view.GravityCompat;
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarDrawerToggle;
@@ -96,6 +97,8 @@ public class DevicesActivity extends AppCompatActivity
             {
                 nav_id = String.valueOf(item.getItemId()); //NOTE, We get the item id here
 
+                getSupportFragmentManager().popBackStack(null, FragmentManager.POP_BACK_STACK_INCLUSIVE);
+
                 Fragment fragment = GenericFillButtonFragment.newInstance(nav_id, "Favourites");
                 HelperMethods.startFragmentMethod(this, fragment, false);
                 Log.d(TAG, "Favourite Tab created");
@@ -103,6 +106,9 @@ public class DevicesActivity extends AppCompatActivity
             }
             case R.id.navigation_rooms:
             {
+                //TODO, Clear backstack here
+                getSupportFragmentManager().popBackStack(null, FragmentManager.POP_BACK_STACK_INCLUSIVE);
+
                 ListRoomFragment listRoomFragment = new ListRoomFragment();
                 HelperMethods.startFragmentMethod(this, listRoomFragment, false);
                 return true;