+++ SpruceCube/app/src/main/java/com/sprucecube/homeautomation/misc/Params.java
@@ -3,6 +3,9 @@ package com.sprucecube.homeautomation.misc;
 
 public class Params
 {
+    //fragmentStack
+    public static final String FAV_STACK = "fav_stack";
+    public static final String BACK_STACK = "back_stack";
 
     //genericFragment
     public static final String NAVIGATION_ID = "navigation_id";
@@ -14,19 +17,17 @@ public class Params
     public static final String CONTROL_ACTION = "Control";
     public static final String PIN_NUM = "pin_num";
 
-
     public static final String INTENT_BUNDLE = "intent_bundle";
 
     //Preferences
     public static final String PREFS = "my_prefs";
     public static final String SETTINGS = "my_settings";
+    public static final String TAG_NAME = "name";
 
     //Filenames
     public static final String ROOM_FILE = "roomfile.txt";
     public static final String ROOM_IMAGE_FILE = "roomimagefile.txt";
 
-    //Fragment Manager
-    public static final String BACK_STACK = "back_stack";
 
     //For AddRoomActivity
     public static final String UPDATE_BOOL = "update_bool";